// src/types.ts
export interface Attachment {
  type: 'image' | 'location' | 'split' | 'emoji' | 'mentions' | 'reply';
  url?: string;
  name?: string;
  lat?: string;
  lng?: string;
  token?: string;
  placeholder?: string;
  charmap?: number[][];
  user_ids?: string[];
  loci?: number[][];
  reply_id?: string;
}

export interface Message {
  id: string;
  source_guid: string;
  created_at: number;
  user_id: string;
  group_id: string;
  name: string;
  avatar_url: string | null;
  text: string | null;
  system: boolean;
  favorited_by: string[];
  attachments: Attachment[];
  sender_type: 'user' | 'bot' | 'system';
  sender_id: string;
}
// ...
export interface DirectMessage {
  id: string;
  source_guid: string;
  recipient_id: string;
  user_id: string;
  created_at: number;
  name: string;
  avatar_url: string | null;
  text: string | null;
  favorited_by: string[];
  attachments: Attachment[];
  sender_id: string;
  sender_type: string;
  conversation_id: string;
}